requirements.txt
langchain
streamlit
python-dotenv
langchain_mistralai
mistralai